create CosmosDBMetricsProvider
