+++ src/WebJobs.Extensions.CosmosDB/Trigger/CosmosDBScaleMonitor.cs
@@ -5,7 +5,6 @@ using System;
 using System.Collections.Generic;
 using System.Diagnostics;
 using System.Linq;
-using System.Net;
 using System.Threading.Tasks;
 using Microsoft.Azure.Cosmos;
 using Microsoft.Azure.WebJobs.Host.Scale;
@@ -18,101 +17,19 @@ namespace Microsoft.Azure.WebJobs.Extensions.CosmosDB.Trigger
         private readonly ILogger _logger;
         private readonly string _functionId;
         private readonly Container _monitoredContainer;
-        private readonly Container _leaseContainer;
-        private readonly string _processorName;
         private readonly ScaleMonitorDescriptor _scaleMonitorDescriptor;
-
-        private static readonly Dictionary<string, string> KnownDocumentClientErrors = new Dictionary<string, string>()
-        {
-            { "Resource Not Found", "Please check that the CosmosDB container and leases container exist and are listed correctly in Functions config files." },
-            { "The input authorization token can't serve the request", string.Empty },
-            { "The MAC signature found in the HTTP request is not the same", string.Empty },
-            { "Service is currently unavailable.", string.Empty },
-            { "Entity with the specified id does not exist in the system.", string.Empty },
-            { "Subscription owning the database account is disabled.", string.Empty },
-            { "Request rate is large", string.Empty },
-            { "PartitionKey value must be supplied for this operation.", "We do not support lease containers with partitions at this time. Please create a new lease collection without partitions." },
-            { "The remote name could not be resolved:", string.Empty },
-            { "Owner resource does not exist", string.Empty },
-            { "The specified document collection is invalid", string.Empty }
-        };
+        private readonly CosmosDBMetricsProvider _cosmosDBMetricsProvider;
 
         public CosmosDBScaleMonitor(string functionId, ILoggerFactory loggerFactory, Container monitoredContainer, Container leaseContainer, string processorName)
         {
             _logger = loggerFactory.CreateLogger<CosmosDBScaleMonitor>();
             _functionId = functionId;
             _monitoredContainer = monitoredContainer;
-            _leaseContainer = leaseContainer;
-            _processorName = processorName;
             _scaleMonitorDescriptor = new ScaleMonitorDescriptor($"{_functionId}-CosmosDBTrigger-{_monitoredContainer.Database.Id}-{_monitoredContainer.Id}".ToLower());
+            _cosmosDBMetricsProvider = new CosmosDBMetricsProvider(loggerFactory, monitoredContainer, leaseContainer, processorName);
         }
 
         public ScaleMonitorDescriptor Descriptor => _scaleMonitorDescriptor;
-
-        public async Task<CosmosDBTriggerMetrics> GetMetricsAsync()
-        {
-            int partitionCount = 0;
-            long remainingWork = 0;
-
-            try
-            {
-                List<ChangeFeedProcessorState> partitionWorkList = new List<ChangeFeedProcessorState>();
-                ChangeFeedEstimator estimator = _monitoredContainer.GetChangeFeedEstimator(_processorName, _leaseContainer);
-                using (FeedIterator<ChangeFeedProcessorState> iterator = estimator.GetCurrentStateIterator())
-                {
-                    while (iterator.HasMoreResults)
-                    {
-                        FeedResponse<ChangeFeedProcessorState> response = await iterator.ReadNextAsync();
-                        partitionWorkList.AddRange(response);
-                    }
-                }
-
-                partitionCount = partitionWorkList.Count;
-                remainingWork = partitionWorkList.Sum(item => item.EstimatedLag);
-            }
-            catch (Exception e) when (e is CosmosException || e is InvalidOperationException)
-            {
-                if (!TryHandleCosmosException(e))
-                {
-                    _logger.LogWarning(Events.OnScaling, "Unable to handle {0}: {1}", e.GetType().ToString(), e.Message);
-                    if (e is InvalidOperationException)
-                    {
-                        throw;
-                    }
-                }
-            }
-            catch (System.Net.Http.HttpRequestException e)
-            {
-                string errormsg;
-
-                var webException = e.InnerException as WebException;
-                if (webException != null &&
-                    webException.Status == WebExceptionStatus.ProtocolError)
-                {
-                    string statusCode = ((HttpWebResponse)webException.Response).StatusCode.ToString();
-                    string statusDesc = ((HttpWebResponse)webException.Response).StatusDescription;
-                    errormsg = string.Format("CosmosDBTrigger status {0}: {1}.", statusCode, statusDesc);
-                }
-                else if (webException != null &&
-                    webException.Status == WebExceptionStatus.NameResolutionFailure)
-                {
-                    errormsg = string.Format("CosmosDBTrigger Exception message: {0}.", webException.Message);
-                }
-                else
-                {
-                    errormsg = e.ToString();
-                }
-
-                _logger.LogWarning(Events.OnScaling, errormsg);
-            }
-
-            return new CosmosDBTriggerMetrics
-            {
-                Timestamp = DateTime.UtcNow,
-                PartitionCount = partitionCount,
-                RemainingWork = remainingWork
-            };
-        }
 
         public ScaleStatus GetScaleStatus(ScaleStatusContext<CosmosDBTriggerMetrics> context)
         {
@@ -121,7 +38,12 @@ namespace Microsoft.Azure.WebJobs.Extensions.CosmosDB.Trigger
 
         async Task<ScaleMetrics> IScaleMonitor.GetMetricsAsync()
         {
-            return await GetMetricsAsync();
+            return await GetMetricsAsync().ConfigureAwait(false);
+        }
+
+        public async Task<CosmosDBTriggerMetrics> GetMetricsAsync()
+        {
+            return await _cosmosDBMetricsProvider.GetMetricsAsync().ConfigureAwait(false);
         }
 
         public ScaleStatus GetScaleStatus(ScaleStatusContext context)
@@ -220,32 +142,6 @@ namespace Microsoft.Azure.WebJobs.Extensions.CosmosDB.Trigger
             _logger.LogInformation(Events.OnScaling, $"CosmosDB container '{_monitoredContainer.Id}' is steady.");
 
             return status;
-        }
-
-        // Since all exceptions in the Cosmos client are thrown as CosmosExceptions, we have to parse their error strings because we dont have access to the internal types
-        private bool TryHandleCosmosException(Exception exception)
-        {
-            string errormsg = null;
-            string exceptionMessage = exception.Message;
-
-            if (!string.IsNullOrEmpty(exceptionMessage))
-            {
-                foreach (KeyValuePair<string, string> exceptionString in KnownDocumentClientErrors)
-                {
-                    if (exceptionMessage.IndexOf(exceptionString.Key, StringComparison.OrdinalIgnoreCase) >= 0)
-                    {
-                        errormsg = !string.IsNullOrEmpty(exceptionString.Value) ? exceptionString.Value : exceptionMessage;
-                    }
-                }
-            }
-
-            if (!string.IsNullOrEmpty(errormsg))
-            {
-                _logger.LogWarning(errormsg);
-                return true;
-            }
-
-            return false;
         }
 
         private static bool IsTrueForLast(IList<CosmosDBTriggerMetrics> metrics, int count, Func<CosmosDBTriggerMetrics, CosmosDBTriggerMetrics, bool> predicate)
